requirements.txt
# Handles the asynchronous internet communication for the Binance library
aiohttp>=3.11.0

# Provides the security tools needed to sign and encrypt your API requests
cryptography>=44.0.0

# Manages the recording of all bot actions and errors into the log file
loguru>=0.7.2

# Automatically reads and validates the settings and keys from the .env file
pydantic-settings>=2.8.0

# The main framework used to write and run the automated test suite
pytest>=8.3.0

# Plugin that allows pytest to run the asynchronous trade functions
pytest-asyncio>=0.25.0

# Plugin that calculates the percentage of your code covered by tests
pytest-cov>=6.0.0

# The primary library used to interact with the Binance Futures API
python-binance>=1.0.21

# Helper tool to load variables from the .env file into the system memory
python-dotenv>=1.0.1

# Standard library for making basic web requests if needed for health checks
requests>=2.32.0

# The framework used to build the interactive terminal menu and prompts
typer>=0.15.0